Moved all the MIDI related logic from INotationPlayback into IMasterNotationMidiData
