+++ src/notation/CMakeLists.txt
@@ -41,6 +41,7 @@ set(MODULE_SRC
     ${CMAKE_CURRENT_LIST_DIR}/inotationstyle.h
     ${CMAKE_CURRENT_LIST_DIR}/inotationaccessibility.h
     ${CMAKE_CURRENT_LIST_DIR}/inotationmidiinput.h
+    ${CMAKE_CURRENT_LIST_DIR}/imasternotationmididata.h
     ${CMAKE_CURRENT_LIST_DIR}/notationtypes.h
     ${CMAKE_CURRENT_LIST_DIR}/inotationconfiguration.h
     ${CMAKE_CURRENT_LIST_DIR}/notationerrors.h
@@ -95,9 +96,8 @@ set(MODULE_SRC
     ${CMAKE_CURRENT_LIST_DIR}/internal/searchcommandsparser.cpp
     ${CMAKE_CURRENT_LIST_DIR}/internal/searchcommandsparser.h
     ${CMAKE_CURRENT_LIST_DIR}/internal/inotationselectionrange.h
-    ${CMAKE_CURRENT_LIST_DIR}/internal/inotationmidievents.h
-    ${CMAKE_CURRENT_LIST_DIR}/internal/notationmidievents.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/internal/notationmidievents.h
+    ${CMAKE_CURRENT_LIST_DIR}/internal/masternotationmididata.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/internal/masternotationmididata.h
     ${CMAKE_CURRENT_LIST_DIR}/internal/instrumentsrepository.cpp
     ${CMAKE_CURRENT_LIST_DIR}/internal/instrumentsrepository.h
