+++ src/notation/inotationplayback.h
@@ -23,6 +23,7 @@
 #define MU_NOTATION_INOTATIONPLAYBACK_H
 
 #include <QRect>
+
 #include "retval.h"
 #include "midi/miditypes.h"
 
@@ -32,15 +33,7 @@ namespace mu::notation {
 class INotationPlayback
 {
 public:
-
-    using InstrumentTrackId = std::string;
-
     virtual ~INotationPlayback() = default;
-
-    virtual std::vector<InstrumentTrackId> instrumentTrackIdList() const = 0;
-    virtual midi::MidiData instrumentMidiData(const InstrumentTrackId& id) const = 0;
-    virtual async::Channel<InstrumentTrackId> instrumentTrackRemoved() const = 0;
-    virtual async::Channel<InstrumentTrackId> instrumentTrackAdded() const = 0;
 
     virtual QTime totalPlayTime() const = 0;
 
@@ -50,8 +43,6 @@ public:
     virtual QRect playbackCursorRectByTick(midi::tick_t tick) const = 0;
 
     virtual RetVal<midi::tick_t> playPositionTickByElement(const Element* element) const = 0;
-
-    virtual Ret playElementMidiData(const Element* element) = 0;
 
     enum BoundaryTick : midi::tick_t {
         FirstScoreTick = 0,
