+++ src/notation/internal/notationplayback.h
@@ -29,8 +29,6 @@
 
 #include "../inotationplayback.h"
 #include "igetscore.h"
-#include "inotationmidievents.h"
-#include "inotationparts.h"
 #include "inotationconfiguration.h"
 
 namespace Ms {
@@ -45,16 +43,9 @@ class NotationPlayback : public INotationPlayback, public async::Asyncable
     INJECT(notation, INotationConfiguration, configuration)
 
 public:
-    NotationPlayback(IGetScore* getScore, async::Notification notationChanged, INotationMidiEventsPtr midiDataProvider);
-    ~NotationPlayback();
+    NotationPlayback(IGetScore* getScore, async::Notification notationChanged);
 
-    void init(INotationPartsPtr parts);
-
-    void load();
-    midi::MidiData instrumentMidiData(const InstrumentTrackId& id) const override;
-    std::vector<InstrumentTrackId> instrumentTrackIdList() const override;
-    async::Channel<InstrumentTrackId> instrumentTrackRemoved() const override;
-    async::Channel<InstrumentTrackId> instrumentTrackAdded() const override;
+    void init();
 
     QTime totalPlayTime() const override;
 
@@ -64,8 +55,6 @@ public:
     QRect playbackCursorRectByTick(midi::tick_t tick) const override;
 
     RetVal<midi::tick_t> playPositionTickByElement(const Element* element) const override;
-
-    Ret playElementMidiData(const Element* element) override;
 
     void addLoopBoundary(LoopBoundaryType boundaryType, midi::tick_t tick) override;
     void setLoopBoundariesVisible(bool visible) override;
@@ -77,16 +66,6 @@ public:
 
 private:
     Ms::Score* score() const;
-    Ms::MasterScore* masterScore() const;
-
-    midi::TempoMap makeTempoMap() const;
-
-    int instrumentBank(const Ms::Instrument* instrument) const;
-
-    // play element
-    Ret playNoteMidiData(const Ms::Note* note) const;
-    Ret playChordMidiData(const Ms::Chord* chord) const;
-    Ret playHarmonyMidiData(const Ms::Harmony* harmony) const;
 
     void addLoopIn(int tick);
     void addLoopOut(int tick);
@@ -95,19 +74,9 @@ private:
 
     const Ms::TempoText* tempoText(int tick) const;
 
-    midi::MidiData buildMidiData(const Ms::Part* part) const;
-    midi::MidiMapping buildMidiMapping(const Ms::Part* part) const;
-    midi::MidiStream buildMidiStream(const Ms::Part* part) const;
-
     IGetScore* m_getScore = nullptr;
-    INotationPartsPtr m_notationParts = nullptr;
-    INotationMidiEventsPtr m_midiEventsProvider = nullptr;
     async::Channel<int> m_playPositionTickChanged;
     ValCh<LoopBoundaries> m_loopBoundaries;
-
-    std::unordered_map<InstrumentTrackId, midi::MidiData> m_instrumentsMidiData;
-    async::Channel<InstrumentTrackId> m_instrumentTrackRemoved;
-    async::Channel<InstrumentTrackId> m_instrumentTrackAdded;
 };
 }
